requirements.txt
gcastle
pandas
numpy
matplotlib
networkx
torch